# Refactoring and performance improvement.