## Before

In [1]: %timeit conf(821112, 28062010)
1000 loops, best of 3: **640 µs per loop**

In [2]: %timeit conv(821112, 28062010)
1 loops, best of 3: **3.32 s per loop**

## After

In [3]: %timeit continued_fraction(821112, 28062010)
100000 loops, best of 3: **4.5 µs per loop**

In [4]: %timeit convergents(continued_fraction(821112, 28062010))
1000 loops, best of 3: **908 µs per loop**